added support for a faux parent
The use case would be when faking a product into a category for which is does not actually live within. Ex: single product in multiple categories. By allowing the passing of a fauxParentId you can divert the breadcrumb from its original family tree and into another. Result is a breadcrumb path that matches the category you are presenting the product (resource) to be and not the actual path it resides in.